whisper-server/requirements.txt
flask
flask-cors
openai-whisper
yt-dlp
psycopg2-binary
google-genai
python-dotenv
